# go-swagger/cmd/api/api-docs/swagger.yaml
basePath: /
definitions:
  core.ErrorMsg:
    properties:
      message:
        type: string
    type: object
  core.ProductCreateReq:
    properties:
      category_name:
        type: string
      desc:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    type: object
  domain.Product:
    properties:
      category_name:
        type: string
      desc:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    type: object
host: localhost:8090
info:
  contact: {}
  description: Example project demonstrating Swagger with net/http.
  title: My API with net/http
  version: "1.0"
paths:
  /products:
    get:
      description: Returns a list of products. If the optional 'category' header is
        provided, the result will be filtered by that category.
      parameters:
      - description: Filter by category
        in: header
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Product'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorMsg'
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a product with the input payload
      parameters:
      - description: Create product
        in: body
        name: productCreateReq
        required: true
        schema:
          $ref: '#/definitions/core.ProductCreateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorMsg'
      summary: Create a product
      tags:
      - products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorMsg'
      summary: Delete a product
      tags:
      - products
    get:
      consumes:
      - application/json
      description: get a product with id param
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorMsg'
      summary: Get a product
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update an existing product by its ID using the provided JSON body
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/domain.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorMsg'
      summary: Update product
      tags:
      - products
  /products/{productId}/img:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image for a product
      parameters:
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorMsg'
      summary: Upload product image
      tags:
      - products
swagger: "2.0"